+++ src/main/webapp/components/MultiTab.js
@@ -35,7 +35,6 @@ export const MultiTab = hh(class MultiTab extends Component {
           onSelect: this.handleSelect,
         }, [
           this.props.children.map((child, idx) => {
-            console.log("child.props.isRendered => ", child.props);
             return h(
               Tab,
               { className: "tabContent", key: idx, eventKey: child.key, title: child.props.title, isRendered: isEmpty(child.props.hide) ? true : !child.props.hide },
