+++ src/main/webapp/qaReport/QaReport.js
@@ -10,164 +10,24 @@ import { QA_EVENT_SORT_NAME_INDEX } from "../util/ReportConstants";
 import { MultiTab } from "../components/MultiTab";
 import { Panel } from "../components/Panel";
 import { datesDiff } from "../util/Utils";
+import IrbTable from "./IrbTable";
+import NoIrbTable from "./NoIrbTable";
 
-const QA_REPORT_SPINNER = "qaReportSpinner";
-const columns = [{
-  dataField: 'id',
-  text: 'Id',
-  hidden: true,
-  csvExport : false
-}, {
-  dataField: 'projectKey',
-  text: 'Project',
-  sort: true
-}, {
-  dataField: 'type',
-  text: 'Type',
-  sort: true
-}, {
-  dataField: 'status',
-  text: 'Status',
-  sort: true
-}, {
-  dataField: 'age',
-  text: 'Age',
-  sort: true,
-  formatter: (cell, row, rowIndex, colIndex) => {
-    const result = datesDiff(new Date(), new Date(row.requestDate));
-    return  (result.years > 0 ? result.years + ' years, ' : '') +
-            (result.months > 0 ? result.months + ' months ' : '') +
-            (result.months > 0 || result.years > 0 ? ' and ': '') + result.days + ' days'
-  }
-}, {
-  dataField: 'actor',
-  text: 'Assignees',
-  sort: true
-}];
-
-const SIZE_PER_PAGE_LIST = [
-  { text: '50', value: 50 },
-  { text: '100', value: 100 },
-  { text: '500', value: 500 }];
-
-const defaultSorted = [{
-  dataField: 'date',
-  order: 'desc'
-}];
+export const QA_REPORT_SPINNER = "qaReportSpinner";
 
 class QaReport extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      sizePerPage: 50,
-      search: null,
-      sort: {
-        sortDirection: 'asc',
-        orderColumn: null
-      },
-      currentPage: 1,
-      recordsTotal: 0,
-      recordsFiltered: 0,
       activeTab: 'IRB',
-      data: [],
+      applyFilter: false,
       beforeDate: null,
       afterDate: null,
       projectType: { value: 'all', label: 'All' },
       showError: false,
-      isAdmin: component.isAdmin,
-      hideIrbTab: false,
-      hideNoIrbTab: false
+      isAdmin: component.isAdmin
     };
   }
-
-  async componentDidMount() {
-    await this.init();
-  }
-
-  async init() {
-    await this.tableHandler(0, this.state.sizePerPage, this.state.search, this.state.sort, this.state.currentPage, this.state.activeTab);
-  }
-
-  tableHandler = async (offset, limit, search, sort, page, tab) => {
-    let query = {
-      draw: 1,
-      start: offset,
-      length: limit,
-      orderColumn: sort.orderColumn,
-      sortDirection: sort.sortDirection,
-      searchValue: search,
-    };
-
-    let filter = {
-      before: this.state.beforeDate ? this.state.beforeDate.toLocaleDateString('en-US') : this.state.afterDate,
-      after: this.state.afterDate ? this.state.afterDate.toLocaleDateString('en-US') : this.state.afterDate,
-      projectType: this.state.projectType.value
-    };
-    console.log("Before ", filter.before, "\nAfter ", filter.after);
-    spinnerService.show(QA_REPORT_SPINNER);
-    let result = await Reports.getQaEventReport(query, tab, filter);
-
-    const lastPage = Math.ceil(result.data.recordsFiltered / query.length);
-    this.setState(prev => {
-      prev.activeTab = tab;
-      prev.lastPage = lastPage;
-      prev.currentPage = page;
-      prev.data = result.data.data;
-      prev.recordsTotal = result.data.recordsTotal;
-      prev.recordsFiltered = result.data.recordsFiltered;
-      prev.sizePerPage = query.length;
-      prev.search = query.searchValue;
-      prev.sort = {
-        orderColumn : query.orderColumn,
-        sortDirection: query.sortDirection
-      };
-      return prev;
-    }, () => {
-      spinnerService.hide(QA_REPORT_SPINNER)
-    })
-    // }).catch(error => {
-    //   spinnerService.hide(QA_REPORT_SPINNER);
-    //   this.setState(() => { throw error });
-    // });
-  };
-
-  onTableChange = async (type, newState) => {
-    switch(type) {
-      case TABLE_ACTIONS.SEARCH: {
-        await this.onSearchChange(newState.searchText);
-        break
-      }
-      case TABLE_ACTIONS.FILTER: {
-        // Not implemented
-        break;
-      }
-      case TABLE_ACTIONS.PAGINATION: {
-        await this.onPageChange(newState.page, newState.sizePerPage);
-        break;
-      }
-      case TABLE_ACTIONS.SORT: {
-        await this.onSortChange(newState.sortField, newState.sortOrder);
-        break;
-      }
-    }
-  };
-
-  onSearchChange = async (search) => {
-    await this.tableHandler(0, this.state.sizePerPage, search, this.state.sort, 1, this.state.activeTab);
-  };
-
-  onPageChange = async (page, sizePerPage) => {
-    const offset = (page - 1) * sizePerPage;
-    await this.tableHandler(offset, sizePerPage, this.state.search, this.state.sort, page, this.state.activeTab);
-  };
-
-  onSortChange = async (sortName, sortOrder) => {
-    const sort = {
-      sortDirection: sortOrder,
-      orderColumn: QA_EVENT_SORT_NAME_INDEX[sortName]
-    };
-    this.tableHandler(0, this.state.sizePerPage, null, sort, this.state.activeTab)
-  };
 
   handleDatePicker = (e) => (date) => {
     this.setState(prev => {
@@ -183,28 +43,24 @@ class QaReport extends Component {
     }) ;
   };
 
-  tabChange = async (tab) => {
-    await this.tableHandler(0, this.state.sizePerPage, this.state.search, this.state.sort, 1, tab)
-  };
-
   applyFilterPanel = async () => {
     this.setState(prev => {
-      prev.hideIrbTab = !(this.state.projectType.value === 'irb' || this.state.projectType.value === 'all');
-      prev.hideNoIrbTab = !(this.state.projectType.value !== 'irb' || this.state.projectType.value === 'all');
+      prev.applyFilter = true;
       return prev;
-    }
-  );
-    await this.tableHandler(0, this.state.sizePerPage, this.state.search, this.state.sort, 1, this.state.activeTab)
+    });
+    // await this.tableHandler(0, this.state.sizePerPage, this.state.search, this.state.sort, 1, this.state.activeTab)
   };
 
   clearFilterPanel = () => {
     this.setState(prev => {
-      prev.hideIrbTab = false;
-      prev.hideNoIrbTab = false;
       prev.beforeDate = null;
       prev.afterDate = null;
       prev.projectType = { value: 'all', label: 'All' };
-    }, async () => await this.tableHandler(0, this.state.sizePerPage, this.state.search, this.state.sort, 1, 'IRB'));
+      prev.applyFilter = false;
+      return prev;
+    }
+    // , async () => await this.tableHandler(0, this.state.sizePerPage, this.state.search, this.state.sort, 1, 'IRB')
+    );
   };
 
   render() {
@@ -221,42 +77,38 @@ class QaReport extends Component {
           applyFilterPanel: this.applyFilterPanel,
           clearFilterPanel: this.clearFilterPanel
         }),
-        Panel({
-          title:
+        div({ className: "headerBoxContainer" }, [
+          div({ className: "containerBox" }, [
             MultiTab({
               defaultActive: this.state.activeTab,
-              handleSelect: (tab) => this.tabChange(tab)
+              handleSelect: (tab) => console.log(tab)
             }, [
               div({
                 key: "IRB",
                 title: "IRB Projects",
-                hide: this.state.hideIrbTab
-              }, []),
+                hide: false
+              }, [
+                h(IrbTable, {
+                  beforeDate: this.state.beforeDate,
+                  afterDate: this.state.afterDate,
+                  projectType: this.state.projectType,
+                  applyFilter: this.state.applyFilter
+                })
+              ]),
               div({
                 key: "NO_IRB",
                 title: "Non IRB Projects",
-                hide: this.state.hideNoIrbTab
-              }, []),
+                hide: false
+              }, [
+                h(NoIrbTable, {
+                  beforeDate: this.state.beforeDate,
+                  afterDate: this.state.afterDate,
+                  projectType: this.state.projectType,
+                  applyFilter: this.state.applyFilter
+                })
+              ])
             ])
-        },[
-          TableComponent({
-            remoteProp: true,
-            onTableChange: this.onTableChange,
-            data: this.state.data,
-            columns: columns,
-            keyField: 'id',
-            search: true,
-            fileName: 'ORSP',
-            page: this.state.currentPage,
-            totalSize: this.state.recordsFiltered,
-            showPrintButton: true,
-            sizePerPageList: SIZE_PER_PAGE_LIST,
-            printComments: this.printComments,
-            defaultSorted: defaultSorted,
-            pagination: true,
-            showExportButtons: true,
-            showSearchBar: true
-          })
+          ])
         ]),
         h(Spinner, {
           name: QA_REPORT_SPINNER, group: "orsp", loadingImage: component.loadingImage
@@ -264,6 +116,5 @@ class QaReport extends Component {
       ])
     );
   }
-
 }
 export default QaReport;
